MainMenuSceneImplementation
A Main Menu scene was created
(The template was based on the Samsung Note 20 Ultra 5G)

Start, Exit, and Settings buttons were added.

A background was added.

A Title with a template for the name was added.

Custom fonts for Legacy Text were uploaded.

Placeholders for scripts on the buttons and title were added.